Prevents package query error to show up on logs.
As we changed the way we query for installed packages with dpkg, the
error that it reports when the package is installed is showing up on the
logs.

Instead we suppress the error message to prevent the following messages:

```
tools >> No packages found matching git-svn. Installing...
tools >> Running "apt-get install -y git-svn"
Package `ack-grep' is not installed and no info is available.
Use dpkg --info (= dpkg-deb --info) to examine archive files,
and dpkg --contents (= dpkg-deb --contents) to list their contents.
```